# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TECA API",
    "description": "\nAPI del e-commerce de muebles **TECA — Para toda la vida**.\n\nCubre la tienda pública (catálogo, carrito, checkout de invitado o con cuenta),\nel área del cliente (pedidos, direcciones, métodos de pago, devoluciones) y el\npanel de administración (usuarios internos, productos, pedidos, cupones,\nauditoría y finanzas).\n\n### Autenticación\nTodas las rutas protegidas usan **JWT Bearer**:\n\n1. `POST /auth/login` con correo y contraseña.\n2. Copia el `access_token` de la respuesta.\n3. Pulsa **Authorize** (arriba a la derecha) y pégalo.\n\n### Roles\n| Rol | Alcance |\n|---|---|\n| `cliente` | Tienda: carrito, pedidos propios, cuenta, reseñas |\n| `admin` | Acceso total al panel |\n| `editor` | Productos |\n| `encargado` | Productos, pedidos y devoluciones |\n| `vendedor` | Pedidos (y lectura de productos) |\n| `soporte` | Devoluciones y lectura de pedidos |\n| `finanzas` | Reportes financieros |\n\n### Convenciones\n- Moneda: balboas (B/.), equivalente a USD.\n- Los identificadores son `ObjectId` de Mongo expuestos como `id` (string).\n- Errores en formato `{\"detail\": \"mensaje\"}` con el código HTTP correspondiente.\n",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Registrar un cliente nuevo",
        "description": "Crea una cuenta de cliente con rol `cliente` y correo **sin verificar**.\n\nGenera un token de verificación que en producción se enviaría por correo; con `DEV_MODE=true` se devuelve en la respuesta (`verification_token`) para poder probar el flujo sin servidor de correo.\n\n**Errores:** `409` si el correo ya está registrado.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registro de cliente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "El correo ya está registrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-email/{token}": {
      "get": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Verificar el correo electrónico",
        "description": "Marca la cuenta como verificada usando el token recibido en el registro. Corresponde a la pantalla *«Verifica tu correo para continuar»*.\n\n**Errores:** `400` si el token es inválido o ya fue usado.",
        "operationId": "verify_email_auth_verify_email__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Token de verificación recibido al registrarse",
              "title": "Token"
            },
            "description": "Token de verificación recibido al registrarse"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Token de verificación inválido"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Iniciar sesión y obtener el token JWT",
        "description": "Valida correo y contraseña y devuelve:\n\n- `access_token`: JWT para el header `Authorization: Bearer <token>`.\n- `must_change_password`: `true` cuando es un usuario interno con contraseña temporal, para que el frontend lo obligue a cambiarla en el primer ingreso.\n- `user`: datos básicos (id, nombre, correo, rol, correo verificado).\n\nSirve tanto para clientes como para usuarios internos del panel; el rol viene dentro del token y en `user.role`.\n\n**Errores:** `401` credenciales incorrectas · `403` cuenta inactiva.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inicio de sesión"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Correo o contraseña incorrectos"
          },
          "403": {
            "description": "La cuenta está inactiva"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Solicitar recuperación de contraseña",
        "description": "Pantalla *«¿Olvidaste tu contraseña?»*. Genera un token de reseteo válido por **1 hora**.\n\nPor seguridad responde siempre el mismo mensaje, exista o no el correo, para no revelar qué cuentas están registradas. Con `DEV_MODE=true` se incluye el `reset_token` en la respuesta para poder probar el flujo.",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitud de recuperación"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Restablecer la contraseña con el token",
        "description": "Define una contraseña nueva (mínimo 8 caracteres) usando el token enviado por correo. El token se invalida tras usarse.\n\n**Errores:** `400` si el token es inválido o pasó más de 1 hora.",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Restablecer contraseña"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Token inválido o expirado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Cambiar la contraseña estando autenticado",
        "description": "Requiere la contraseña actual. También es el endpoint que usa el usuario interno para reemplazar su contraseña temporal (deja `must_change_password` en `false`).\n\n🔒 Requiere token. **Errores:** `400` si la contraseña actual no coincide.",
        "operationId": "change_password_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cambio de contraseña"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "La contraseña actual es incorrecta"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Ver mi perfil",
        "description": "Datos del usuario dueño del token (pantalla *Mi cuenta*). Nunca incluye el hash de la contraseña.\n\n🔒 Requiere token.",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Editar mi perfil",
        "description": "Actualiza nombre y/o teléfono. Los campos omitidos o nulos se dejan intactos (actualización parcial).\n\n🔒 Requiere token.",
        "operationId": "update_profile_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Edición de perfil"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error de validación"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}